inventory 2 finished

+++ app/Http/Controllers/AdminEquipmentController.php
@@ -14,6 +14,7 @@ use Illuminate\Http\Request;
 use App\Http\Requests;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Redirect;
+use Illuminate\View\View;
 
 class AdminEquipmentController extends Controller
 {
@@ -30,14 +31,16 @@ class AdminEquipmentController extends Controller
 
     public function index()
     {
+
+        $equipments = Equipment::all();
+
         $users = User::lists('name', 'id')->all();
         $borrows = Borrow::all();
 
-        $equipments = Equipment::all()->where('status', 1);
 
         $equipmentdrop = Equipment::lists('item', 'id')->all();
 
-        return view('admin.equipments.index', compact('equipments', 'equipmentdrop', 'borrows', 'borrow', 'borrows2', 'users'));
+        return view('admin.equipments.index', compact('equipments', 'equipmentdrop', 'borrows', 'borrow', 'users', 'findusers'));
 
     }
 
@@ -60,52 +63,72 @@ class AdminEquipmentController extends Controller
      */
     public function store(Request $request)
     {
-//        dd($request);
+
+
         if (User::find($request->input('name'))) {
             $users = User::find($request->input('name'));
 
             $borrow = new Borrow([
-                'name' => $users->name,
-                'description' => $users->email,
                 'user_id' => Auth::user()->id,
+                'borrowedby_id'=> $users->id
             ]);
 
-            foreach ($request->originalQuantity as $key => $value) {
-                if (array_key_exists($key, $request->quantity) && array_key_exists($key, $request->originalQuantity) && array_key_exists($key, $request->borrows))
-                    $item = NonConsumable::create(['quantity' => ($request->originalQuantity[$key] - $request->quantity[$key]), 'item' => $request->borrows[$key], 'status' => 0]);
-                $quantityBorrow = NonConsumable::create(['quantity' => $request->quantity[$key], 'item' => $request->borrows[$key], 'status' => 1]);
-                $itemid = $item->id;
-                $quantityBorrowId[] = $quantityBorrow->id;
 
-//                dd($request->originalQuantity[$key] - $request->quantity[$key]);
+            if ($request->originalQuantity) {
+                foreach ($request->originalQuantity as $key => $value) {
+                    $nonconsumableItem = Equipment::findOrfail($request->borrows[$key]);
 
-                if (!($request->originalQuantity[$key] - $request->quantity[$key]) == 0) {
-                    foreach ($request->borrows as $id) {
-                        Equipment::where('id', $id)->update(['status' => 1, 'nonconsumable_id' => $itemid]);
-                    }
+                    if (array_key_exists($key, $request->quantity) && array_key_exists($key, $request->originalQuantity) && array_key_exists($key, $request->borrows))
+                        $item = NonConsumable::create(['quantity' => ($request->originalQuantity[$key] - $request->quantity[$key]), 'item' => $request->borrows[$key], 'status' => 0, 'outOfStock' => 0]);
 
-                }else{
-                    foreach ($request->borrows as $id) {
-                        Equipment::where('id', $id)->update(['status' => 0, 'nonconsumable_id' => $itemid]);
-                    }
+                    $quantityBorrow = NonConsumable::create(['quantity' => $request->quantity[$key], 'item' => $request->borrows[$key], 'status' => 1]);
+                    $itemid = $item->id;
                     
+                    $quantityBorrowId[] = $quantityBorrow->id;
+
+                    if (!($request->originalQuantity[$key] - $request->quantity[$key] <= 0)) {
+
+                        Equipment::where('id', $request->borrows[$key])->update(['status' => 1, 'nonconsumable_id' => $itemid, 'consumable' => 1]);
+
+                    } else {
+
+                        if (($request->originalQuantity[$key] - $request->quantity[$key]) > 0) {
+                            Equipment::where('id', $request->borrows[$key])->update(['status' => 0, 'nonconsumable_id' => $itemid, 'consumable' => 1]);
+                        } else {
+                            Equipment::where('id', $request->borrows[$key])->update(['status' => 0, 'nonconsumable_id' => $itemid, 'consumable' => 1, 'outOfStock' => 1, 'hasQuantity' => 0]);
+                        }
+                    }
+
+                    foreach ($quantityBorrowId as $borrowId) {
+                        NonConsumable::where('id', $borrowId)->update(['status' => 1, 'name' => $nonconsumableItem->item]);
+                        $borrow->nonconsumable_id = $borrowId;
+
+                    }
+
+
                 }
 
-                foreach ($quantityBorrowId as $borrowId) {
-                    NonConsumable::where('id', $borrowId)->update(['status' => 1, 'currentquantity' => $request->originalQuantity[$key] - $request->quantity[$key] ]);
+            } else {
+
+                foreach ($request->non as $key => $value) {
+                    NonConsumable::create(['quantity' => -1, 'item' => $request->non[$key], 'status' => 0,]);
+                    foreach ($request->non as $id) {
+                        Equipment::where('id', $id)->update(['status' => 0, 'nonconsumable_id' => $request->non[$key], 'consumable' => 0]);
+                    }
                 }
 
             }
 
-//            $quantityBorrowId
-//                foreach ($request->borrows as $id) {
-//                    Equipment::where('id', $id)->update(['status' => 0]);
-//                }
-
-
             $borrow->save();
-            $borrow->equipments()->sync($request->borrows, false);
-            $borrow->nonconsumables()->sync($quantityBorrowId, false);
+
+            if ($request->originalQuantity) {
+                $borrow->equipments()->sync($request->borrows, false);
+                $borrow->names()->sync([$request->input('name')], false);
+                $borrow->nonconsumables()->sync($quantityBorrowId, false);
+            } else {
+                session()->flash('nonconsumables', 'Successful');
+                $borrow->equipments()->sync($request->non, false);
+            }
             session()->flash('success', 'Borrowed successfully save!');
             return Redirect::route('admin.borrow.index');
         } else {
@@ -118,6 +141,7 @@ class AdminEquipmentController extends Controller
                 'photo_id' => 'required',
             ]);
             $input = $request->all();
+
             $user = Auth::user();
             if ($file = $request->file('photo_id')) {
                 $name = time() . $file->getClientOriginalName();
@@ -125,15 +149,20 @@ class AdminEquipmentController extends Controller
                 $photo = Photo::create(['file' => $name]);
                 $input['photo_id'] = $photo->id;
             }
+            $input['consumable'] = 1;
             if ($request->nonconsumable_id) {
                 $quantity = $request->nonconsumable_id;
                 $nonconsumable = NonConsumable::create(['quantity' => $quantity]);
                 $input['nonconsumable_id'] = $nonconsumable->id;
-            }
 
+            } else {
+                session()->flash('borrows', 'The Equipment was successfully save!');
+            }
             $user->equipment()->create($input);
             session()->flash('success', 'The Equipment was successfully save!');
+
             return Redirect::route('admin.equipment.index');
+
         }
 
 
@@ -147,7 +176,7 @@ class AdminEquipmentController extends Controller
      */
     public function show($id)
     {
-        //
+
     }
 
     /**
@@ -160,6 +189,7 @@ class AdminEquipmentController extends Controller
     {
         $categories = Category::lists('name', 'id')->all();
         $equipments = Equipment::findOrfail($id);
+
         return view('admin.equipments.edit', compact('equipments', 'categories'));
     }
 
@@ -172,6 +202,7 @@ class AdminEquipmentController extends Controller
      */
     public function update(Request $request, $id)
     {
+
         $input = $request->all();
         if ($file = $request->file('photo_id')) {
             $name = time() . $file->getClientOriginalName();
@@ -179,6 +210,22 @@ class AdminEquipmentController extends Controller
             $photo = Photo::create(['file' => $name]);
             $input['photo_id'] = $photo->id;
         }
+
+        $input['consumable'] = 1;
+
+        if ($request->nonconsumable_id) {
+
+            $quantity = $request->nonconsumable;
+            NonConsumable::where('id', $request->nonconsumable_id)->update(['quantity' => $quantity]);
+            $input['outOfStock'] = 0;
+        } else {
+            session()->flash('borrows', 'The Equipment was successfully save!');
+        }
+
+
+        session()->flash('success', 'The Equipment was successfully save!');
+
+
         Auth::user()->equipment()->whereId($id)->first()->update($input);
         return redirect(route('admin.equipment.index'));
     }
@@ -191,12 +238,12 @@ class AdminEquipmentController extends Controller
      */
     public function destroy(Request $request, $id = null)
     {
-
         if (is_null($id) == !$request->has('no')) {
             session()->flash('error', 'no record');
             return back();
         } elseif (is_null($id) == $request->has('no')) {
             $str = trim($request->no, ",");
+
             $ids = explode(",", $str);
             $intArray = array_map(
                 function ($value) {
@@ -206,7 +253,9 @@ class AdminEquipmentController extends Controller
             );
             foreach ($intArray as $id) {
                 $equipment = Equipment::findOrFail($id);
-                unlink(public_path() . $equipment->photo->file);
+                if (!($equipment->photo_id == 0)) {
+                    unlink(public_path() . $equipment->photo->file);
+                }
                 $equipment->delete();
             }
         } elseif (!is_null($id)) {
